refactor(generate-g4): extract profile and classification helpers
